add staff authorisation and token authorisation to deleteEvents and createEvents

+++ controllers/event-controllers.ts
@@ -81,7 +81,7 @@ export const createEvent = async (req: Request, res: Response) => {
       uid: userId,
     });
 
-    if (!user) {
+    if (!user || user.role !== "staff") {
       res.status(404).json({ msg: "User not found" });
       return;
     }
@@ -219,6 +219,22 @@ export const deleteEvent = async (req: Request, res: Response) => {
       res.status(400).json({ msg: `Invalid event ID format` });
       return;
     }
+
+    const authToken = extractTokenFromAuthorization(req.headers.authorization!);
+    console.log(authToken);
+
+    const decodeValue = await supabaseClient.auth.getUser(authToken);
+    const userId = decodeValue.data.user?.id;
+
+    const user = await userSchema.findOne({
+      uid: userId,
+    });
+
+    if (!user || user.role !== "staff") {
+      res.status(403).json({ msg: "User is not authorised" });
+      return;
+    }
+
     const deletedEvent = await eventSchema.findByIdAndDelete(id);
 
     if (!deletedEvent) {
